refactor: 디렉터리명 변경
